adding some styling to marker forms and dialog
