+++ src/app/fish-map/form-dialog.html
@@ -8,23 +8,29 @@
 
 <div mat-dialog-content>
   <p mat-dialog-title>{{markerNotes}}</p>
+  <p>Fish</p>
+  
   <h4>Enter Catch Data</h4>
+  <mat-divider></mat-divider>
   <form>
-
     <mat-form-field>
       <input matInput [matDatepicker]="picker1" placeholder="Angular forms" [formControl]="formDate">
       <mat-datepicker-toggle matSuffix [for]="picker1"></mat-datepicker-toggle>
       <mat-datepicker #picker1></mat-datepicker>
     </mat-form-field>
-
-    <div class="form-group">
+    <mat-divider></mat-divider>
+    <mat-form-field>
       <label for="notes">Notes</label>
       <textarea matInput minRows="10" [(ngModel)]="catchModel.notes" class="form-control" id="notes" name="notes"></textarea>
-    </div>
-
+    </mat-form-field>
+    <mat-divider></mat-divider>
+    <mat-form-field>
+      <label for="notes">Number of Fish</label>
+      <input matInput type="number" minRows="10" [(ngModel)]="catchModel.fish.number" class="form-control" id="notes" name="notes">
+    </mat-form-field>
     <button type="button" (click)="onSubmitClick()" class="btn btn-success">Submit</button>
-
   </form>
+
 </div>
 
 <div>
